resume-chatbot-backend/requirements.txt
fastapi
uvicorn
langchain==0.1.16
langchain-community==0.0.34
langchain-groq==0.1.3
faiss-cpu
pypdf
python-dotenv
sentence-transformers
